api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
deprecation==2.0.7
distlib==0.3.6
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.12.2
frozenlist==1.3.3
gpt-engineer==0.0.7
h11==0.14.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
nodeenv==1.8.0
openai==0.27.8
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
pydantic==1.10.11
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
requests==2.31.0
rudder-sdk-python==2.0.2
ruff==0.0.272
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
virtualenv==20.23.1
yarl==1.9.2
